ml_service/requirements.txt
apache-airflow[celery,postgres,mongo]==2.10.3
pandas
pymongo
joblib
matplotlib
fastapi
uvicorn[standard]
u8darts[torch]
